move json schema logic into LS, improve schema loading
